Shell: Add get return code command.
This new commmand, and corresponding update to command
execution, enables the user to retrieve the return code
of the most recently run shell command. This is useful both
for users wanting specific information beyond the print
statements in a command, but especially for automated users
that prefer a simple return code to parsing human readable text.
This was tested using all default shell commands, as well as
eeprom, flash, and i2c, where specific errors could be forced.
In all cases, the correct return value was returned by the new
retval command.

Signed-off-by: Hunter Searle <hsearle@xes-inc.com>