usb: host: add class API
Add class driver data and node structures, as well as an instantiation
macro similar to those found in the USB device stack. Update the class
driver API accordingly. The idea is to use iterable sections to detect
the appropriate driver based on the code triple and, potentially, the
VID/PID.  While it is not currently possible to create a device at
runtime, the class driver should be designed to handle this in future
and to avoid misuse of the DT for instantiation.

Signed-off-by: Johann Fischer <johann.fischer@nordicsemi.no>